% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LogisticModelDevelopment.R
\name{LogisticModelDevelopment}
\alias{LogisticModelDevelopment}
\title{Logistic Model Development}
\usage{
LogisticModelDevelopment(var)
}
\arguments{
\item{See}{the example}

\item{Note-}{please download woe package from GitHub using "install_github('riv','tomasgreif')". After installing the correct package if the UI closes automatically please restart the rstudio.}
}
\value{
you can download the model and all the performance
}
\description{
Give you ui to create logistic model.
}
\examples{
m <-data.frame()
x <- sapply(sapply(ls(), get), is.data.frame)
m <- data.frame(names(x)[(x == TRUE)])
LogisticModelDevelopment(m)
}
\author{
JIshnu
}
